backend/requirements.txt
fastapi 
uvicorn[standard] 
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
